+++ ChoGGi's Library/Code/OnMsgs.lua
@@ -1,3 +1,6 @@
+local StringFormat = string.format
+local S = ChoGGi.Strings
+
 --~ ChoGGi.Temp.UIScale = (LocalStorage.Options.UIScale + 0.0) / 100 or 100
 ChoGGi.Temp.UIScale = (LocalStorage.Options.UIScale + 0.0) / 100
 -- used for resizing my dialogs to scale
@@ -14,5 +17,25 @@ end
 function OnMsg.ReloadLua()
 	if type(XShortcutsTarget.UpdateToolbar) == "function" then
 		ChoGGi.ComFuncs.Rebuildshortcuts()
+	end
+end
+
+-- use this message to perform post-built actions on the final classes
+function OnMsg.ClassesBuilt()
+	-- add cat for my items
+	local bc = BuildCategories
+	if not table.find(bc,"id","ChoGGi") then
+		local image = StringFormat("%sUI/bmc_incal_resources.png",ChoGGi.LibraryPath)
+		local highlight = StringFormat("%sUI/bmc_incal_resources_shine.png",ChoGGi.LibraryPath)
+
+		bc[#bc+1] = {
+			id = "ChoGGi",
+			name = S[302535920001400--[[ChoGGi--]]],
+			image = image,
+			highlight = highlight,
+			-- pre-gagarin
+			img = image,
+			highlight_img = highlight,
+		}
 	end
 end