+++ Mods ChoGGi/RC Garage/metadata.lua
@@ -1,7 +1,7 @@
 return PlaceObj("ModDef", {
-	"title", "RC Garage v0.1",
-	"version", 1,
-	"saved", 1541419200,
+	"title", "RC Garage v0.2",
+	"version", 2,
+	"saved", 1541505600,
 	"image", "Preview.png",
 	"id", "ChoGGi_RCGarage",
 	"steam_id", "1557866331",